requirements.txt
PyYAML==6.0.1
yt-dlp==2023.11.16
